课程信息持久化框架搭建初步

+++ DSCourseDesign/src/main/java/mapper/StudentMapper.java
@@ -5,5 +5,5 @@ import org.apache.ibatis.annotations.Param;
 
 public interface StudentMapper {
 
-     Student select(@Param("id") long id, @Param("password") String password);
+     Student select(@Param("id") int id, @Param("password") String password);
 }
